feature(#4782): Refactor export endpoints

See merge request canopsis/canopsis-pro!2273

+++ community/go-engines-community/lib/api/entity/models.go
@@ -61,7 +61,7 @@ type ExportResponse struct {
 	//   * `0` - Running
 	//   * `1` - Succeeded
 	//   * `2` - Failed
-	Status int `json:"status"`
+	Status int64 `json:"status"`
 }
 
 type Entity struct {
